Fix build errors and warnings

+++ presentation/src/main/kotlin/pl/karol202/bphelper/presentation/viewmodel/MembersViewModel.kt
@@ -13,5 +13,5 @@ interface MembersViewModel : Closeable
 
 	fun updateMember(member: MemberViewData)
 
-	fun removeMember(memberId: Int)
+	fun removeMember(memberId: Long)
 }
